# DBRedis: use HMGET instead of HGET