Redis HMGET request allows querying multiple fields at the same time, reducing time spent in round trip on network, as well as system resource usage when Redis server is local.

Use it with a maximum of DB_REDIS_HMGET_NUMKEYS fields per request. Its value (30) was chosen to keep most responses in a single IP packet when running on the same machine as Redis server. It should not be set too high since Redis is single threaded and players could suffer latency problems.