@Nestorfish 2016-10-19
Any comment?

Regarding performance, here is a bench with a map stored on local Redis server:
Before:
```
$ time build/minetestmapper -i ../minetest/worlds/test/ -o test.png --drawalpha 
Warning: Falling back to using colors.txt from current directory.
Warning: The width or height of the image to be created exceeds 4096 pixels! (Dimensions: 5136x5136)
      
real	3m16,582s
user	2m6,353s
sys	0m34,077s
```
After:
```
$ time build/minetestmapper -i ../minetest/worlds/test/ -o test.png --drawalpha 
Warning: Falling back to using colors.txt from current directory.
Warning: The width or height of the image to be created exceeds 4096 pixels! (Dimensions: 5136x5136)

real	1m59,274s
user	1m50,457s
sys	0m1,750s
```
I can move the hack for std::string::c_str() validity to the beginning of HMGET() function if it helps understanding, memory overhead should be negligible (<1MiB).

Don't hesitate to ask me about correction or explanation, the performance increase is worth it.